# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test Case API",
        "description": "An API of test case",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1"
        }
    ],
    "paths": {
        "/messages": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Get list of messages",
                "description": "Returns list of messages",
                "operationId": "getMessagesList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "MessageResource": {
                "title": "Message Resource",
                "properties": {
                    "uuid": {
                        "type": "string",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "status": {
                        "type": "string",
                        "example": "Sent"
                    },
                    "content": {
                        "type": "string",
                        "example": "Hello, world!"
                    },
                    "user": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "phone_number": {
                                "type": "string",
                                "example": "+123456789"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Messages",
            "description": "Messages"
        }
    ]
}